Make eclipse java type inference happy

+++ src/test/java/org/liveontologies/puli/RandomProofNode.java
@@ -37,7 +37,7 @@ public class RandomProofNode {
 		for (int i = 0; i < maxInferences; i++) {
 			int conclusion = random.nextInt(maxConclusions);
 			 ProofBuilder<Integer, ?, ?>.ProofInferenceBuilder ib = b
-					.conclusion(conclusion);
+					 .conclusion((Integer) conclusion);
 			int noPremises = random.nextInt(maxPremises);
 			if (derived.size() < noPremises) {
 				noPremises = derived.size();
